+++ MyBank/src/test/java/com/revature/repository/CustomerRepoTest.java
@@ -43,18 +43,7 @@ public class CustomerRepoTest {
 
              assertFalse(mockCustomerRepo.addCustomer(customer));
     }
-    @Test
-    public void updateAccount() {
 
-        mockCustomerRepo.UpdateAccount("1000878",4000.00);
-    }
-
-    @Test
-    public void addCustomerAccount() {
-
-        mockCustomerRepo.addCustomerAccount(customer);
-
-    }
 
     @Test
     public void findUserByUsername() {
@@ -64,17 +53,4 @@ public class CustomerRepoTest {
 
     }
 
-    @Test
-    public void findUserByAccountNo() {
-        assertTrue(customerRepo.findAccountByAccountNo(1000697).isPresent());
-        assertEquals("Gu",customerRepo.findUserByAccountNo(100005).get().getLastName());
-
-
-    }
-
-    @Test
-    public void findAccountByAccountNo() {
-        assertTrue(customerRepo.findAccountByAccountNo(100005).isPresent());
-        assertEquals("100005",customerRepo.findAccountByAccountNo(100005).get().getAccountNo());
-    }
 }
